zynqmp_pmufw: Fix violations for MISRA Rule 10_3
This patch fix following violation
misra_c_2012_rule_10_3_violation: The value of an expression shall
not be assigned to an object with a narrower essential type or
of a different essential type category.

To solve this MISRA violation do explicit type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_clock.c
@@ -578,7 +578,7 @@ static s32 PmClockGenGetGateState(PmClock* const clock, u8* const enable)
 	s32 status = PmClockGateGetShift(clk, &shift);
 
 	if (XST_SUCCESS == status) {
-		*enable = (XPfw_Read32(clk->ctrlAddr) >> shift) & 1U;
+		*enable = (u8)((XPfw_Read32(clk->ctrlAddr) >> shift) & 1U);
 	}
 
 	return status;
@@ -2785,7 +2785,7 @@ void PmClockRelease(PmNode* const node)
 		PmClockReleaseInt(&ch->clock->base);
 		ch = ch->nextClock;
 	}
-	node->flags &= ~NODE_LOCKED_CLOCK_FLAG;
+	node->flags &= ~(u8)NODE_LOCKED_CLOCK_FLAG;
 
 done:
 	return;
