+++ lib/sw_apps/zynqmp_pmufw/src/pm_system.c
@@ -513,10 +513,10 @@ s32 PmSystemRequirementAdd(void)
 			goto done;
 		}
 		req->info |= PM_SYSTEM_USING_SLAVE_MASK;
-		req->preReq = pmSystemReqs[i].caps;
-		req->currReq = pmSystemReqs[i].caps;
-		req->nextReq = pmSystemReqs[i].caps;
-		req->defaultReq = pmSystemReqs[i].caps;
+		req->preReq = (u8)pmSystemReqs[i].caps;
+		req->currReq = (u8)pmSystemReqs[i].caps;
+		req->nextReq = (u8)pmSystemReqs[i].caps;
+		req->defaultReq = (u8)pmSystemReqs[i].caps;
 		req->latencyReq = MAX_LATENCY;
 	}
 
@@ -593,7 +593,7 @@ void PmSystemRestartDone(const PmMaster* const master)
 	req = PmRequirementGetNoMaster(&pmSlavePl_g);
 	caps = PmSystemGetRequirement(&pmSlavePl_g);
 	if ((NULL != req) && (0U == (PM_CAP_ACCESS & caps))) {
-		req->currReq &= ~PM_CAP_ACCESS;
+		req->currReq &= ~(u8)PM_CAP_ACCESS;
 	}
 
 	return;
